+++ dzgui/init/migrate.py
@@ -1,5 +1,4 @@
 import shutil
-import sys
 
 from pathlib import Path
 from dzgui.const.constants import LEGACY_CONFIG_PATH, LEGACY_COLS_PATH, LEGACY_IPS_PATH
@@ -7,16 +6,11 @@ from dzgui.config.convert import rc2json
 from dzgui.util._json import read_json, write_json
 
 
-# TODO: move to setup wizard
 def migrate_legacy_conf(config: Path) -> None:
     old_conf = Path.home() / LEGACY_CONFIG_PATH
-    if old_conf.is_file():
-        j = rc2json(old_conf)
-        config.parent.mkdir(parents=True, exist_ok=True)
-        config.write_text(j)
-    else:
-        print("Unimplemented. You must have a working dztuirc.")
-        sys.exit(1)
+    j = rc2json(old_conf)
+    config.parent.mkdir(parents=True, exist_ok=True)
+    config.write_text(j)
 
 
 def has_new_config(config: Path) -> bool:
@@ -24,6 +18,7 @@ def has_new_config(config: Path) -> bool:
 
 
 def migrate_cols_file(res: Path) -> None:
+    # NOTE: dzg.columns.json is API 7 spec
     old_res = Path.home() / LEGACY_COLS_PATH
     if old_res.is_file():
         j = read_json(old_res)
@@ -38,11 +33,20 @@ def migrate_cols_file(res: Path) -> None:
 def copy_state_files(state_path: Path) -> None:
     home = Path.home()
     legacy = home / ".local/state/dzgui"
+    to_copy = [
+        "dzg.res.json",
+        "dzg.notes.json",
+        "dzg.history",
+        "dzg.versions",
+        "ips.csv",
+        ".month",
+    ]
     if state_path == legacy:
         # TODO: log this
         return
     for file in legacy.iterdir():
-        shutil.copy(file, state_path / file.name)
+        if file.name in to_copy:
+            shutil.copy(file, state_path / file.name)
 
 
 def copy_ipdb(ips_path: Path) -> None: