+++ dzgui/init/update.py
@@ -1,14 +1,10 @@
 import logging
 import requests
-import subprocess
-import sys
 
-from importlib import resources
 from packaging.version import Version
 
-from dzgui.const.constants import APP_NAME, APP_NAME_LOWER, REQUEST_TIMEOUT
+from dzgui.const.constants import APP_NAME, REQUEST_TIMEOUT
 from dzgui.const.endpoints import GITHUB_RELEASES, CODEBERG_RELEASES
-from dzgui.init.prefix import is_prefix_writeable
 
 logger = logging.getLogger(APP_NAME)
 
@@ -28,31 +24,13 @@ def get_latest_release() -> str | None:
     return tag
 
 
-def allow_updates(allow: bool) -> bool:
-    if allow is False:
-        return False
-    if allow is True:
-        return is_prefix_writeable()
-
-
-def check_updates(version: str) -> None:
+def check_updates(version: str) -> bool:
     try:
         latest = get_latest_release()
-        prefix = sys.prefix
         if latest is None:
-            return
+            return False
         if Version(version) >= Version(latest):
-            return
-
-        # TODO: test update logic
-        print("UNIMPLEMENTED: fetches in-app updates")
-        return
-
-        with resources.path(APP_NAME_LOWER, "scripts/update.sh") as path:
-            proc = subprocess.Popen(["/usr/bin/env", "bash", path, latest, prefix])
-            if proc != 0:
-                # TODO: pop a dialog
-                pass
-            sys.exit(proc)
+            return False
+        return True
     except Exception:
-        return
+        return False