+++ dzgui/main.py
@@ -13,29 +13,27 @@ from dzgui.config.ipdb import get_ipdb
 from dzgui.config.query import lookup
 from dzgui.config.userprefs import UserPrefs
 from dzgui.config.xdg import get_xdg_paths, parse_filepaths
-from dzgui.const.update import ALLOW_UPDATES
 from dzgui.init.coords import get_local_coords
 from dzgui.init.dayz import is_dayz_installed
 from dzgui.init.flock import lock_acquire
 from dzgui.init.migrate import (
     has_new_config,
-    migrate_cols_file,
-    migrate_legacy_conf,
     copy_state_files,
+    migrate_cols_file,
 )
 from dzgui.init.prefix import get_version
 from dzgui.init.prereqs import has_steam_client
-from dzgui.init.update import allow_updates, check_updates
-
 from dzgui.strings import boot
+from dzgui.init.update import check_updates
 
-from dzgui.util.map_count import get_map_count
+# from dzgui.util.map_count import get_map_count
 from dzgui.util.deck import is_steam_deck, is_game_mode
 from dzgui.util.localize import set_locale
 from dzgui.util.symlink import rebuild_symlinks
 from dzgui.util.strings import init, flags
 
 from dzgui.views.base import App
+from dzgui.views.dialogs.wizard import SetupWizard
 from dzgui.views.dialogs.early_alert import EarlyAlertDialog
 
 if TYPE_CHECKING:
@@ -97,20 +95,9 @@ def main() -> None:
 
     if XDG.resolution.parent.is_dir() is False:
         make_parents(XDG.resolution)
-    # TODO: test
+
     if XDG.debug.is_file() is False:
         make_parents(XDG.debug)
-
-    if has_new_config(XDG.config) is False:
-        # TODO: handle this in assistant
-        migrate_legacy_conf(XDG.config)
-        migrate_cols_file(XDG.columns)
-        # TODO: copy notes file
-        copy_state_files(xdg_paths["XDG_STATE_HOME"])
-
-    setup_logger(XDG.debug)
-    with open(XDG.debug, "w") as f:
-        f.truncate(0)
 
     _is_steam_deck = is_steam_deck()
     _is_game_mode = is_game_mode() if _is_steam_deck else False
@@ -118,12 +105,18 @@ def main() -> None:
         # NOTE: this may no longer be necessary on newer versions of SteamOS
         del os.environ["GTK_IM_MODULE"]
 
-    # TODO: test spamming timeout
-    allow = allow_updates(ALLOW_UPDATES)
-    if allow is True:
-        check_updates(version)
+    if has_new_config(XDG.config) is False:
+        migrate_cols_file(XDG.columns)
+        copy_state_files(xdg_paths["XDG_STATE_HOME"])
+        # TODO: add logging inside wizard
+        SetupWizard(_is_steam_deck, XDG.config)
+        return
 
-    # TODO: config wizard: check has_steam_client() prior to VDF exploration
+    setup_logger(XDG.debug)
+    with open(XDG.debug, "w") as f:
+        f.truncate(0)
+
+    update_available = check_updates(version)
 
     if _is_steam_deck is False:
         # TODO: sudo escalation dialog
@@ -148,10 +141,9 @@ def main() -> None:
         is_debug=args.debug,
         coords=local_coords,
         version=version,
-        allow_updates=allow,
         paths=XDG,
+        update_available=update_available,
         use_miles=use_miles,
     )
-    # TODO: drop allow updates
     print(boot.all_ok)
     App(prefs)