+++ dzgui/views/components/buttons.py
@@ -21,7 +21,7 @@ from dzgui.const.constants import (
 import gi
 
 gi.require_version("Gtk", "3.0")
-from gi.repository import Gtk, GLib  # noqa E402
+from gi.repository import Gtk, Gdk, GLib  # noqa E402
 
 if TYPE_CHECKING:
     from gi.repository import GLib